+++ src/dev/scene_editor.rs
@@ -1,6 +1,6 @@
 use crate::file_system_interaction::game_serialization::{GameLoadRequest, GameSaveRequest};
 use crate::file_system_interaction::level_serialization::{WorldLoadRequest, WorldSaveRequest};
-use crate::level_instanciation::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
+use crate::level_instantiation::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
 use crate::movement::navigation::navmesh::NavMesh;
 use crate::player_control::actions::{Actions, ActionsFrozen};
 use crate::GameState;
