+++ src/file_system_interaction/game_serialization.rs
@@ -1,5 +1,5 @@
 use crate::file_system_interaction::level_serialization::{CurrentLevel, WorldLoadRequest};
-use crate::level_instanciation::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
+use crate::level_instantiation::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
 use crate::player_control::player_embodiment::Player;
 use crate::world_interaction::condition::ActiveConditions;
 use crate::world_interaction::dialog::{CurrentDialog, DialogEvent};
