+++ src/file_system_interaction/level_serialization.rs
@@ -1,5 +1,5 @@
 use crate::file_system_interaction::asset_loading::LevelAssets;
-use crate::level_instanciation::spawning::{GameObject, SpawnEvent, SpawnTracker};
+use crate::level_instantiation::spawning::{GameObject, SpawnEvent, SpawnTracker};
 use crate::world_interaction::condition::ActiveConditions;
 use crate::world_interaction::dialog::CurrentDialog;
 use bevy::prelude::*;
