+++ src/level_instantiation/spawning/objects/player.rs
@@ -1,10 +1,8 @@
-use crate::level_instanciation::spawning::post_spawn_modification::CustomCollider;
-use crate::level_instanciation::spawning::{
+use crate::level_instantiation::spawning::{
     GameObject, PrimedGameObjectSpawner, PrimedGameObjectSpawnerImplementor,
 };
 use crate::movement::general_movement::{CharacterAnimations, KinematicCharacterBundle, Model};
-use crate::movement::navigation::Follower;
-use crate::world_interaction::dialog::{DialogId, DialogTarget};
+use crate::player_control::player_embodiment::{Player, PlayerSensor};
 use bevy::prelude::*;
 use bevy_rapier3d::prelude::*;
 use std::f32::consts::TAU;
@@ -12,9 +10,9 @@ use std::f32::consts::TAU;
 pub const HEIGHT: f32 = 1.;
 pub const RADIUS: f32 = 0.4;
 
-pub struct NpcSpawner;
+pub struct PlayerSpawner;
 
-impl PrimedGameObjectSpawnerImplementor for NpcSpawner {
+impl PrimedGameObjectSpawnerImplementor for PlayerSpawner {
     fn spawn<'a, 'b: 'a>(
         &self,
         spawner: &'b mut PrimedGameObjectSpawner<'_, '_, 'a>,
@@ -24,7 +22,7 @@ impl PrimedGameObjectSpawnerImplementor for NpcSpawner {
         let gltf = spawner
             .gltf
             .get(&spawner.scenes.character)
-            .unwrap_or_else(|| panic!("Failed to load scene for NPC"));
+            .unwrap_or_else(|| panic!("Failed to load scene for player"));
 
         spawner
             .commands
@@ -33,9 +31,9 @@ impl PrimedGameObjectSpawnerImplementor for NpcSpawner {
                     transform,
                     ..default()
                 },
-                Name::new("NPC"),
+                Player,
+                Name::new("Player"),
                 KinematicCharacterBundle::capsule(HEIGHT, RADIUS),
-                Follower,
                 CharacterAnimations {
                     idle: spawner.animations.character_idle.clone(),
                     walk: spawner.animations.character_walking.clone(),
@@ -44,28 +42,24 @@ impl PrimedGameObjectSpawnerImplementor for NpcSpawner {
             ))
             .with_children(|parent| {
                 parent.spawn((
-                    DialogTarget {
-                        dialog_id: DialogId::new("follower"),
-                    },
-                    Name::new("NPC Dialog Collider"),
-                    Collider::cylinder(HEIGHT / 2., RADIUS * 5.),
+                    Collider::capsule_y(HEIGHT / 2., RADIUS),
                     Sensor,
-                    ActiveEvents::COLLISION_EVENTS,
-                    ActiveCollisionTypes::KINEMATIC_STATIC,
-                    CustomCollider,
+                    PlayerSensor,
+                    ActiveCollisionTypes::all(),
+                    Name::new("Player Sensor"),
                 ));
                 parent.spawn((
                     SceneBundle {
                         scene: gltf.scenes[0].clone(),
                         transform: Transform {
                             translation: Vec3::new(0., -HEIGHT, 0.),
-                            scale: Vec3::splat(0.012),
                             rotation: Quat::from_rotation_y(TAU / 2.),
+                            scale: Vec3::splat(0.01),
                         },
                         ..default()
                     },
                     Model,
-                    Name::new("NPC Model"),
+                    Name::new("Player Model"),
                 ));
             })
             .id()
